requirements.txt
torch==1.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.62.3
numpy==1.20.3
matplotlib==3.4.3
pillow==8.4.0
pandas==1.3.4